refactor(app): move the users routes into a (commun) group
