+++ apps/app/app/(authed)/collectivite/[collectiviteId]/(commun)/layout.tsx
